pyproject.toml
[tool.poetry]
name = "camda2021-dialogi"
version = "0.1.0"
description = ""
authors = ["Nicholas M Katritsis <nmk37@cam.ac.uk>"]

[tool.poetry.dependencies]
python = "~3.9"
scipy = "^1.7.3"
pandas = "^1.3.5"
stanza = "^1.3.0"
Unidecode = "^1.3.2"
flatten-dict = "^0.4.2"
psycopg2-binary = "^2.9.3"
h5py = "^3.6.0"
faiss-gpu = "^1.7.2"
tensorflow = "^2.7.0"
scikit-learn = "^1.0.2"
xmltodict = "^0.12.0"
networkx = "^2.6.3"
node2vec = "^0.4.4"
matplotlib = "^3.5.1"
keras-tuner = "^1.1.0"
tensorflow-addons = "^0.15.0"
seaborn = "^0.11.2"
signaturizer = "^1.1.11"
rdkit-pypi = "^2021.9.4"
rbo = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
